+++ Sources/vapor/cmd_clean.swift
@@ -1,0 +1,17 @@
+
+struct Clean: Command {
+    static let id = "clean"
+    static func execute(with args: [String], in directory: String) {
+        guard args.isEmpty else {
+            fail("\(id) doesn't take any additional parameters")
+        }
+
+        do {
+            try run("rm -rf Packages .build")
+            print("Cleaned.")
+        } catch {
+            fail("Could not clean.")
+        }
+    }
+}
+
